localize string; move NewNotesButton to its own file
