+++ Nos/Assets/Localization/Localizable.xcstrings
@@ -6895,6 +6895,17 @@
         }
       }
     },
+    "newNotesAvailable" : {
+      "extractionState" : "manual",
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "New notes available"
+          }
+        }
+      }
+    },
     "next" : {
       "extractionState" : "manual",
       "localizations" : {
