+++ Nos/Views/Home/HomeFeedView.swift
@@ -158,33 +158,6 @@ struct HomeFeedView: View {
     }
 }
 
-struct NewNotesButton: View {
-    @FetchRequest var newNotes: FetchedResults<Event>
-    var action: (() async -> Void)?
-
-    init(user: Author, lastRefreshDate: Date, seenOn: Relay?, action: @escaping () async -> Void) {
-        let request = Event.homeFeed(for: user, after: lastRefreshDate, seenOn: seenOn)
-        _newNotes = FetchRequest(fetchRequest: request)
-        self.action = action
-    }
-
-    var body: some View {
-        if newNotes.isEmpty {
-            EmptyView()
-        } else {
-            VStack {
-                SecondaryActionButton(
-                    title: "New notes available",
-                    font: .clarity(.semibold, textStyle: .footnote),
-                    action: action
-                )
-                Spacer()
-            }
-            .padding(8)
-        }
-    }
-}
-
 #Preview {
     var previewData = PreviewData()
